Demo breakup 1 (#1146)
* Wallet & CLI Create DLC Offer

* Wallet & CLI Accept DLC Offer

* Wallet & CLI Sign DLC

* Wallet & CLI Add DLC Sigs

+++ dlc/src/main/scala/org/bitcoins/dlc/testgen/DLCTestVectorGenerator.scala
@@ -134,7 +134,7 @@ object DLCTestVectorGenerator {
     )
     val remoteChangeSPK = P2WPKHWitnessSPKV0(ECPublicKey.freshPublicKey)
 
-    val penaltyTimeout = 30
+    val penaltyTimeout = UInt32(30)
     val contractMaturity = BlockTime(UInt32(System.currentTimeMillis() / 1000))
     val contractTimeout = BlockTime(
       contractMaturity.toUInt32 + UInt32(60 * 60 * 24))
